+++ packages/jest-editor-support/src/__tests__/runner.test.js
@@ -183,8 +183,8 @@ describe('Runner', () => {
     it('calls createProcess with the --coverage arg when provided', () => {
       const expected = '--coverage';
 
-      const workspace: any = {};
-      const options = {coverage: true};
+      const workspace: any = {collectCoverage: true};
+      const options = {};
       const sut = new Runner(workspace, options);
       sut.start(false);
 
@@ -196,8 +196,8 @@ describe('Runner', () => {
     it('calls createProcess with the ---no-coverage arg when provided and false', () => {
       const expected = '--no-coverage';
 
-      const workspace: any = {};
-      const options = {coverage: false};
+      const workspace: any = {collectCoverage: false};
+      const options = {};
       const sut = new Runner(workspace, options);
       sut.start(false);