+++ packages/jest-editor-support/src/types.js
@@ -20,7 +20,6 @@ import type {ChildProcess} from 'child_process';
 import type ProjectWorkspace from './project_workspace';
 
 export type Options = {
-  coverage?: boolean,
   createProcess?: (
     workspace: ProjectWorkspace,
     args: Array<string>,